+++ snakemake/deployment/conda.py
@@ -16,7 +16,6 @@ import subprocess
 import tempfile
 import hashlib
 import shutil
-from distutils.version import StrictVersion
 import json
 from glob import glob
 import tarfile
@@ -123,7 +122,7 @@ class Env:
             from snakemake.shell import shell
 
             content = shell.check_output(
-                "conda env export {}".format(self.address_argument),
+                f"conda env export {self.address_argument}",
                 stderr=subprocess.STDOUT,
                 text=True,
             )
@@ -248,9 +247,9 @@ class Env:
     @property
     def address_argument(self):
         if self.is_named:
-            return "--name '{}'".format(self.address)
+            return f"--name '{self.address}'"
         else:
-            return "--prefix '{}'".format(self.address)
+            return f"--prefix '{self.address}'"
 
     @property
     def archive_file(self):
@@ -282,7 +281,7 @@ class Env:
             os.makedirs(env_archive, exist_ok=True)
             try:
                 out = shell.check_output(
-                    "conda list --explicit {}".format(self.address_argument),
+                    f"conda list --explicit {self.address_argument}",
                     stderr=subprocess.STDOUT,
                     text=True,
                 )
@@ -311,14 +310,14 @@ class Env:
                                 tarfile.open(pkg_path)
                         except:
                             raise WorkflowError(
-                                "Package is invalid tar/zip archive: {}".format(pkg_url)
+                                f"Package is invalid tar/zip archive: {pkg_url}"
                             )
         except (
             requests.exceptions.ChunkedEncodingError,
             requests.exceptions.HTTPError,
         ) as e:
             shutil.rmtree(env_archive)
-            raise WorkflowError("Error downloading conda package {}.".format(pkg_url))
+            raise WorkflowError(f"Error downloading conda package {pkg_url}.")
         except (Exception, BaseException) as e:
             shutil.rmtree(env_archive)
             raise e
@@ -401,7 +400,7 @@ class Env:
                     shell.check_output(
                         singularity.shellcmd(
                             self._container_img.path,
-                            "[ -d '{}' ]".format(env_path),
+                            f"[ -d '{env_path}' ]",
                             args=self._singularity_args,
                             envvars=self.get_singularity_envvars(),
                             quiet=True,
@@ -449,9 +448,7 @@ class Env:
                     )
                 )
                 return env_path
-            logger.info(
-                "Creating conda environment {}...".format(self.file.simplify_path())
-            )
+            logger.info(f"Creating conda environment {self.file.simplify_path()}...")
             env_archive = self.archive_file
             try:
                 # Touch "start" flag file
@@ -483,7 +480,7 @@ class Env:
                             "--quiet",
                             "--no-shortcuts" if ON_WINDOWS else "",
                             "--yes",
-                            "--prefix '{}'".format(env_path),
+                            f"--prefix '{env_path}'",
                         ]
                         + packages
                     )
@@ -525,8 +522,8 @@ class Env:
                             + [
                                 "create",
                                 "--quiet",
-                                '--file "{}"'.format(target_env_file),
-                                '--prefix "{}"'.format(env_path),
+                                f'--file "{target_env_file}"',
+                                f'--prefix "{env_path}"',
                             ]
                             + yes_flag
                         )
@@ -609,7 +606,7 @@ class Env:
 
     @classmethod
     def get_singularity_envvars(self):
-        return {"CONDA_PKGS_DIRS": "/tmp/conda/{}".format(uuid.uuid4())}
+        return {"CONDA_PKGS_DIRS": f"/tmp/conda/{uuid.uuid4()}"}
 
     def __hash__(self):
         # this hash is only for object comparison, not for env paths
@@ -653,7 +650,7 @@ class Conda:
             self.frontend = frontend
 
             self.info = json.loads(
-                shell.check_output(self._get_cmd(f"conda info --json"), text=True)
+                shell.check_output(self._get_cmd("conda info --json"), text=True)
             )
 
             if prefix_path is None or container_img is not None:
@@ -740,11 +737,12 @@ class Conda:
             self._check_condarc()
         except subprocess.CalledProcessError as e:
             raise CreateCondaEnvironmentException(
-                f"Unable to check conda installation:" "\n" + e.stderr.decode()
+                "Unable to check conda installation:\n" + e.stderr.decode()
             )
 
     def _check_version(self):
         from snakemake.shell import shell
+        from packaging.version import Version
 
         version = shell.check_output(
             self._get_cmd("conda --version"), stderr=subprocess.PIPE, text=True
@@ -756,9 +754,9 @@ class Conda:
             )
         else:
             version = version_matches[0]
-        if StrictVersion(version) < StrictVersion("4.2"):
+        if Version(version) < Version("4.2"):
             raise CreateCondaEnvironmentException(
-                "Conda must be version 4.2 or later, found version {}.".format(version)
+                f"Conda must be version 4.2 or later, found version {version}."
             )
 
     def _check_condarc(self):
@@ -797,7 +795,7 @@ class Conda:
             activate = activate.replace("\\", "/")
             env_address = env_address.replace("\\", "/")
 
-        return "source {} '{}'; {}".format(activate, env_address, cmd)
+        return f"source {activate} '{env_address}'; {cmd}"
 
     def shellcmd_win(self, env_address, cmd):
         """Prepend the windows activate bat script."""
@@ -805,7 +803,7 @@ class Conda:
         activate = os.path.join(self.bin_path(), "activate.bat").replace("\\", "/")
         env_address = env_address.replace("\\", "/")
 
-        return '"{}" "{}"&&{}'.format(activate, env_address, cmd)
+        return f'"{activate}" "{env_address}"&&{cmd}'
 
 
 def is_mamba_available():